moved Legacy classes and functions into legacy namespace
