+++ src/folf-namespace/prime.cpp
@@ -31,19 +31,3 @@ void folf::prime::primeCalcNew(unsigned long long* maxNum, unsigned long long* p
     }
     *primeSum = primeSumL;
 }
-unsigned long long folf::prime::primeCalcLegacy(unsigned long long* maxNum)
-{
-    unsigned long long primeSum = 2;
-    // OpenMP implementation for multithreading the for loop, completly optional
-    omp_set_num_threads( 9 );                                 // this defines the number of maximal threads, 999 should work fine (threads can be more than actual CPU cores, because most loops are empty)
-    #pragma omp parallel for reduction( + : primeSum )       // if it fails to compile, make sure to set -fopenmp and use a modern compiler shiped with OpenMP
-    for (unsigned long long i = 1; i < *maxNum; i += 2)
-    {
-        // testPrime returns true, if the given number is a Prime
-        if (folf::prime::testPrime(&i) == true)
-        {
-            primeSum = primeSum + i;
-        }
-    }
-    return primeSum;
-}
